pyproject.toml
[build-system]
requires = ['setuptools']
build-backend = "setuptools.build_meta"

[project]
name = 'hello-fastapi'
version = '0.1.0'
readme = 'README.md'

requires-python = '>=3.12'

dependencies = [
    'fastapi == 0.115.6',
    'httpx == 0.28.0',
    # Security
    'pyjwt == 2.10.1',
    'passlib == 1.7.4',
    'bcrypt == 4.2.1',
    # Test
    'pytest == 8.3.4',
    'pytest-cov == 6.0.0',
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = "src"
addopts = "--strict-markers"

[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_also = [
    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self\\.debug",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
    ]

ignore_errors = true